requirements.txt
aiogram==3.22.0
aiohttp==3.9.4
aiosqlite==0.18.0
python-dotenv==1.0.0
pytest>=7.0.0
